Bring in channels view for the logged out home page.
