+++ conferences/package.json
@@ -24,7 +24,8 @@
     "next": "12.0.10",
     "prop-types": "15.8.1",
     "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react-dom": "17.0.2",
+    "react-text-truncate": "0.17.0"
   },
   "devDependencies": {
     "eslint": "8.9.0",
